+++ rasa_core/domain.py
@@ -224,16 +224,19 @@ class Domain(with_metaclass(abc.ABCMeta, object)):
             self.slot_states + \
             self.prev_action_states
 
-    @staticmethod
-    def get_parsing_states(tracker):
+    def get_parsing_states(self, tracker):
         # type: (DialogueStateTracker) -> Dict[Text, float]
 
         state_dict = {}
 
-        # Set all found entities with the state value 1.0
+        # Set all found entities with the state value 1.0, unless they should
+        # be ignored for the current intent
         for entity in tracker.latest_message.entities:
-            key = "entity_{0}".format(entity["entity"])
-            state_dict[key] = 1.0
+            intent_name = tracker.latest_message.intent.get("name")
+            should_use_entity = self._intents[intent_name]['use_entities']
+            if should_use_entity:
+                key = "entity_{0}".format(entity["entity"])
+                state_dict[key] = 1.0
 
         # Set all set slots with the featurization of the stored value
         for key, slot in tracker.slots.items():
@@ -410,8 +413,9 @@ class TemplateDomain(Domain):
             action_factory = data.get("action_factory", None)
         slots = cls.collect_slots(data.get("slots", {}))
         additional_arguments = data.get("config", {})
+        intents = cls.collect_intents(data.get("intents", {}))
         return cls(
-            data.get("intents", []),
+            intents,
             data.get("entities", []),
             slots,
             utter_templates,
@@ -455,6 +459,16 @@ class TemplateDomain(Domain):
             slot = slot_class(slot_name, **slot_dict[slot_name])
             slots.append(slot)
         return slots
+
+    @staticmethod
+    def collect_intents(intent_list):
+        intents = {}
+        for intent in intent_list:
+            if isinstance(intent, dict):
+                intents.update(intent)
+            else:
+                intents.update({intent: {'use_entities': True}})
+        return intents
 
     @staticmethod
     def collect_templates(yml_templates):
@@ -507,10 +521,9 @@ class TemplateDomain(Domain):
         additional_config = {
             "store_entities_as_slots": self.store_entities_as_slots}
         action_names = self.action_names[len(Domain.DEFAULT_ACTIONS):]
-
         domain_data = {
             "config": additional_config,
-            "intents": self.intents,
+            "intents": [{k: v} for k, v in self._intents.items()],
             "entities": self.entities,
             "slots": self._slot_definitions(),
             "templates": self.templates,
@@ -538,7 +551,7 @@ class TemplateDomain(Domain):
 
     @utils.lazyproperty
     def intents(self):
-        return self._intents
+        return sorted(self._intents.keys())
 
     @utils.lazyproperty
     def entities(self):