+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ This project adheres to `Semantic Versioning`_ starting with version 0.2.0.
 Added
 -----
 - new ``RasaChatInput`` channel
+- option to ignore entities for certain intents
 
 Changed
 -------